PR: Refactor how the installers are built

+++ installers-conda/build_conda_pkgs.py
@@ -8,11 +8,9 @@
 import os
 import re
 import sys
-from argparse import ArgumentParser, RawTextHelpFormatter
+from argparse import ArgumentParser
 from configparser import ConfigParser
 from datetime import timedelta
-from logging import Formatter, StreamHandler, getLogger
-from pathlib import Path
 from subprocess import check_call
 from textwrap import dedent
 from time import time
@@ -22,26 +20,19 @@ from git import Repo, rmtree
 from ruamel.yaml import YAML
 from setuptools_scm import get_version
 
-fmt = Formatter('%(asctime)s [%(levelname)s] [%(name)s] -> %(message)s')
-h = StreamHandler()
-h.setFormatter(fmt)
-logger = getLogger('BuildCondaPkgs')
-logger.addHandler(h)
-logger.setLevel('INFO')
+# Local imports
+from utils import logger as logger, HERE, BUILD, DocFormatter
 
-HERE = Path(__file__).parent
-BUILD = HERE / "build"
-RESOURCES = HERE / "resources"
 EXTDEPS = HERE.parent / "external-deps"
-SPECS = BUILD / "specs.yaml"
 REQUIREMENTS = HERE.parent / "requirements"
 REQ_MAIN = REQUIREMENTS / 'main.yml'
 REQ_WINDOWS = REQUIREMENTS / 'windows.yml'
 REQ_MAC = REQUIREMENTS / 'macos.yml'
 REQ_LINUX = REQUIREMENTS / 'linux.yml'
 
-BUILD.mkdir(exist_ok=True)
 SPYPATCHFILE = BUILD / "installers-conda.patch"
+PARENT_BRANCH = os.getenv("MATRIX_BRANCH", os.getenv("GITHUB_BASE_REF"))
+
 
 yaml = YAML()
 yaml.indent(mapping=2, sequence=4, offset=2)
@@ -54,13 +45,9 @@ class BuildCondaPkg:
     source = None
     feedstock = None
     feedstock_branch = None
-    shallow_ver = None
 
-    def __init__(self, data={}, debug=False, shallow=False):
-        self.logger = getLogger(self.__class__.__name__)
-        if not self.logger.handlers:
-            self.logger.addHandler(h)
-        self.logger.setLevel('INFO')
+    def __init__(self, data={}, debug=False):
+        self.logger = logger.getChild(self.__class__.__name__)
 
         self.debug = debug
 
@@ -68,7 +55,7 @@ class BuildCondaPkg:
         self._fdstk_path = BUILD / self.feedstock.split("/")[-1]
         self._patchfile = self._fdstk_path / "recipe" / "version.patch"
 
-        self._get_source(shallow=shallow)
+        self._get_source()
         self._get_version()
 
         self.data = {'version': self.version}
@@ -79,9 +66,10 @@ class BuildCondaPkg:
 
         self._recipe_patched = False
 
-    def _get_source(self, shallow=False):
-        """Clone source and feedstock to distribution directory for building"""
-        self._build_cleanup()
+    def _get_source(self):
+        """Clone source and feedstock for building"""
+        BUILD.mkdir(exist_ok=True)
+        self._cleanup_build()
 
         if self.source == HERE.parent:
             self._bld_src = self.source
@@ -99,12 +87,7 @@ class BuildCondaPkg:
 
             # Clone from source
             kwargs = dict(to_path=self._bld_src)
-            if shallow:
-                kwargs.update(shallow_exclude=self.shallow_ver)
-                self.logger.info(
-                    f"Cloning source shallow from tag {self.shallow_ver}...")
-            else:
-                self.logger.info("Cloning source...")
+            self.logger.info("Cloning source...")
             self.repo = Repo.clone_from(remote, **kwargs)
             self.repo.git.checkout(commit)
 
@@ -113,13 +96,20 @@ class BuildCondaPkg:
         kwargs = dict(to_path=self._fdstk_path)
         if self.feedstock_branch:
             kwargs.update(branch=self.feedstock_branch)
-        Repo.clone_from(self.feedstock, **kwargs)
+        feedstock_repo = Repo.clone_from(self.feedstock, **kwargs)
+        self.logger.info(
+            f"Feedstock branch: {feedstock_repo.active_branch.name}"
+        )
 
-    def _build_cleanup(self):
+    def _cleanup_build(self, debug=False):
         """Remove cloned source and feedstock repositories"""
+        if debug:
+            self.logger.info("Keeping cloned source and feedstock")
+            return
+
         for src in [self._bld_src, self._fdstk_path]:
             if src.exists() and src != HERE.parent:
-                logger.info(f"Removing {src}...")
+                self.logger.info(f"Removing {src}...")
                 rmtree(src)
 
     def _get_version(self):
@@ -234,10 +224,7 @@ class BuildCondaPkg:
             ])
         finally:
             self._recipe_patched = False
-            if self.debug:
-                self.logger.info("Keeping cloned source and feedstock")
-            else:
-                self._build_cleanup()
+            self._cleanup_build(self.debug)
 
             elapse = timedelta(seconds=int(time() - t0))
             self.logger.info(f"Build time = {elapse}")
@@ -248,8 +235,10 @@ class SpyderCondaPkg(BuildCondaPkg):
     norm = False
     source = os.environ.get('SPYDER_SOURCE', HERE.parent)
     feedstock = "https://github.com/conda-forge/spyder-feedstock"
-    feedstock_branch = "main"
-    shallow_ver = "v5.3.2"
+
+    feedstock_branch = "dev"  # Default branch, or if Spyder branch is master
+    if PARENT_BRANCH == "6.x":
+        feedstock_branch = "main"
 
     def _patch_source(self):
         self.logger.info("Patching Spyder source...")
@@ -334,7 +323,6 @@ class PylspCondaPkg(BuildCondaPkg):
     source = os.environ.get('PYTHON_LSP_SERVER_SOURCE')
     feedstock = "https://github.com/conda-forge/python-lsp-server-feedstock"
     feedstock_branch = "main"
-    shallow_ver = "v1.4.1"
 
 
 class QtconsoleCondaPkg(BuildCondaPkg):
@@ -342,15 +330,16 @@ class QtconsoleCondaPkg(BuildCondaPkg):
     source = os.environ.get('QTCONSOLE_SOURCE')
     feedstock = "https://github.com/conda-forge/qtconsole-feedstock"
     feedstock_branch = "main"
-    shallow_ver = "5.3.1"
 
 
 class SpyderKernelsCondaPkg(BuildCondaPkg):
     name = "spyder-kernels"
     source = os.environ.get('SPYDER_KERNELS_SOURCE')
     feedstock = "https://github.com/conda-forge/spyder-kernels-feedstock"
-    feedstock_branch = "rc"
-    shallow_ver = "v2.3.1"
+
+    feedstock_branch = "dev"  # Default branch, or if Spyder branch is master
+    if PARENT_BRANCH == "6.x":
+        feedstock_branch = "main"
 
 
 PKGS = {
@@ -361,17 +350,15 @@ PKGS = {
 }
 
 if __name__ == "__main__":
-    p = ArgumentParser(
-        description=dedent(
+    parser = ArgumentParser(
+        description="Build conda packages to local channel.",
+        epilog=dedent(
             """
-            Build conda packages to local channel.
-
             This module builds conda packages for Spyder and external-deps for
             inclusion in the conda-based installer. The following classes are
             provided for each package:
                 SpyderCondaPkg
                 PylspCondaPkg
-                QdarkstyleCondaPkg
                 QtconsoleCondaPkg
                 SpyderKernelsCondaPkg
 
@@ -385,43 +372,29 @@ if __name__ == "__main__":
             appropriate environment variable from the following:
                 SPYDER_SOURCE
                 PYTHON_LSP_SERVER_SOURCE
-                QDARKSTYLE_SOURCE
                 QTCONSOLE_SOURCE
                 SPYDER_KERNELS_SOURCE
             """
         ),
-        usage="python build_conda_pkgs.py "
-              "[--build BUILD [BUILD] ...] [--debug] [--shallow]",
-        formatter_class=RawTextHelpFormatter
+        formatter_class=DocFormatter
     )
-    p.add_argument(
+    parser.add_argument(
+        '--build', nargs="+", default=list(PKGS.keys()),
+        help="Space-separated list of packages to build."
+    )
+    parser.add_argument(
         '--debug', action='store_true', default=False,
         help="Do not remove cloned sources and feedstocks"
     )
-    p.add_argument(
-        '--build', nargs="+", default=PKGS.keys(),
-        help=("Space-separated list of packages to build. "
-              f"Default is {list(PKGS.keys())}")
-    )
-    p.add_argument(
-        '--shallow', action='store_true', default=False,
-        help="Perform shallow clone for build")
-    args = p.parse_args()
+
+    args = parser.parse_args()
 
     logger.info(f"Building local conda packages {list(args.build)}...")
     t0 = time()
 
-    if SPECS.exists():
-        specs = yaml.load(SPECS.read_text())
-    else:
-        specs = {k: "" for k in PKGS}
-
     for k in args.build:
-        pkg = PKGS[k](debug=args.debug, shallow=args.shallow)
+        pkg = PKGS[k](debug=args.debug)
         pkg.build()
-        specs[k] = "=" + pkg.version
-
-    yaml.dump(specs, SPECS)
 
     elapse = timedelta(seconds=int(time() - t0))
     logger.info(f"Total build time = {elapse}")
